Section 6

+++ src/GymManagement.Application/Subscriptions/Events/SubscriptionDeletedEventHandler.cs
@@ -1,4 +1,3 @@
-using System;
 using GymManagement.Application.Common.Interfaces;
 using GymManagement.Domain.Admins.Events;
 using MediatR;
@@ -6,18 +5,24 @@ using MediatR;
 namespace GymManagement.Application.Subscriptions.Events;
 
 public class SubscriptionDeletedEventHandler(
-  ISubscriptionsRepository subscriptionsRepository,
-  IUnitOfWork unitOfWork) 
-  : INotificationHandler<SubscriptionDeletedEvent>
+    ISubscriptionsRepository subscriptionsRepository,
+    IUnitOfWork unitOfWork)
+        : INotificationHandler<SubscriptionDeletedEvent>
 {
-  private readonly ISubscriptionsRepository _subscriptionsRepository = subscriptionsRepository;
-  private readonly IUnitOfWork _unitOfWork = unitOfWork;
-  public async Task Handle(SubscriptionDeletedEvent notification, CancellationToken cancellationToken)
-  {
-    var subscription = await _subscriptionsRepository.GetByIdAsync(notification.SubscriptionId)
-                        ?? throw new InvalidOperationException();
+    private readonly ISubscriptionsRepository _subscriptionsRepository = subscriptionsRepository;
+    private readonly IUnitOfWork _unitOfWork = unitOfWork;
 
-    await _subscriptionsRepository.RemoveSubscriptionAsync(subscription);
-    await _unitOfWork.CommitChangesAsync();
-  }
+    public async Task Handle(SubscriptionDeletedEvent notification, CancellationToken cancellationToken)
+    {
+        var subscription = await _subscriptionsRepository.GetByIdAsync(notification.SubscriptionId);
+
+        if (subscription is null)
+        {
+            // resilient error handling
+            throw new InvalidOperationException();
+        }
+
+        await _subscriptionsRepository.RemoveSubscriptionAsync(subscription);
+        await _unitOfWork.CommitChangesAsync();
+    }
 }
